# web/backend/static/openapi.yaml
openapi: 3.1.0
info:
  title: ESA Host Brief Viewer
  version: 0.2.0
  description: 'ESA Incentive Pipeline backend API.


    **Authentication**


    - Host endpoints: require `esa_session` cookie (set via `POST /api/login`). Valid
    24h.

    - Admin endpoints: require `esa_admin_session` cookie (set via `POST /api/admin/login`).
    Valid 1h.


    **Base URL**: `https://esa-hosting.codescales.xyz`


    **AI/Tool discovery**


    - OpenAPI spec (this file): `GET /openapi.yaml`

    - Natural-language context doc: `GET /llms.txt`

    - Interactive UI: `GET /docs`'
  contact:
    name: ESA Incentive Pipeline
    url: https://github.com/BongoEADGC6/esa-incentive-pipeline
paths:
...
  /api/notes:
    get:
      tags:
      - notes
      summary: List Notes
      operationId: list_notes_api_notes_get
      parameters:
      - name: run_slug
        in: query
        required: true
        schema:
          type: string
          title: Run Slug
      - name: esa_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Esa Session
      - name: esa_admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Esa Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NoteDTO'
                title: Response List Notes Api Notes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - notes
      summary: Create Note
      operationId: create_note_api_notes_post
      parameters:
      - name: esa_admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Esa Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteCreateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/notes/{note_id}:
    patch:
      tags:
      - notes
      summary: Update Note
      operationId: update_note_api_notes__note_id__patch
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: integer
          title: Note Id
      - name: esa_admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Esa Admin Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - notes
      summary: Delete Note
      operationId: delete_note_api_notes__note_id__delete
      parameters:
      - name: note_id
        in: path
        required: true
        schema:
          type: integer
          title: Note Id
      - name: esa_admin_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Esa Admin Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NoteCreateRequest:
      properties:
        run_slug:
          type: string
          title: Run Slug
        body:
          type: string
          maxLength: 10000
          minLength: 1
          title: Body
        host_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Host Id
      type: object
      required:
      - run_slug
      - body
      title: NoteCreateRequest
    NoteDTO:
      properties:
        id:
          type: integer
          title: Id
        run_slug:
          type: string
          title: Run Slug
        host_id:
          type: integer
          title: Host Id
        host_name:
          type: string
          title: Host Name
        body:
          type: string
          title: Body
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
        is_own:
          type: boolean
          title: Is Own
          default: false
        can_edit:
          type: boolean
          title: Can Edit
          default: false
      type: object
      required:
      - id
      - run_slug
      - host_id
      - host_name
      - body
      - created_at
      - updated_at
      title: NoteDTO
    NoteUpdateRequest:
      properties:
        body:
          type: string
          maxLength: 10000
          minLength: 1
          title: Body
      type: object
      required:
      - body
      title: NoteUpdateRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    hostSession:
      type: apiKey
      in: cookie
      name: esa_session
      description: Set via POST /api/login. Valid 24h.
    adminSession:
      type: apiKey
      in: cookie
      name: esa_admin_session
      description: Set via POST /api/admin/login. Valid 1h.
servers:
- url: https://esa-hosting.codescales.xyz
  description: Production
...